Restructure reporter, add comments
